[REFACTOR] Changed from int to decimal

+++ FinTrac/BusinessLogic/Dtos Components/MovementInXDaysDTO.cs
@@ -3,7 +3,7 @@ namespace BusinessLogic.Dtos_Components
 {
 	public class MovementInXDaysDTO
 	{
-		public int[] Spendings { get; set; }
+		public decimal[] Spendings { get; set; }
